backend/payments/requirements.txt
Django==4.1.7
environs[django]==9.5.0
django-environ==0.10.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
drf-spectacular==0.26.0
pyjwt==2.6.0
django-extensions==3.2.1
yookassa==2.3.5
rollbar==0.16.3
dj-database-url==1.2.0
celery[redis]==5.2.7
django-redis==5.2.0
django-rest-enumfield==0.2.0
gunicorn==20.1.0
gevent==22.10.2
dacite==1.8.0